agentic_ai/agents/mcp_agent_demo/pyproject.toml
[project]
name = "mcp-agent-demo"
version = "0.1.0"
description = "Demo: Expose agent-framework agent as MCP service, consume it from another agent, and orchestrate local+remote agents in a workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-core>=0.2.0",
    "agent-framework-orchestrations>=0.1.0",
    "azure-identity>=1.19.0",
    "fastmcp>=3.0.0",
    "mcp>=1.13.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
]
